requirements.txt
tensorflow

Flask

Flask-Cors

Pillow

numpy

requests

beautifulsoup4

pymongo

selenium